Bundle into package; add sync command and state
