+++ .vscode/settings.json
@@ -1,0 +1,11 @@
+{
+  "[python]": {
+    "editor.insertSpaces": true,
+    "editor.detectIndentation": false,
+    "editor.tabSize": 4,
+    "editor.formatOnSave": true
+  },
+  "python.pythonPath": ".venv/bin/python3",
+  "python.linting.flake8Enabled": true,
+  "python.formatting.provider": "black"
+}
