+++ setup.cfg
@@ -1,0 +1,6 @@
+[flake8]
+max-line-length = 100
+ignore = E203,E266,E501,W503,W504,W605,E402,E302
+max-complexity = 18
+select = B,C,E,F,W,T4,B9
+exclude = .git
